Use Map's insertion order, fix counter overflow
There is no need to count up repeatedly from the `lru` to the `mru`
value.  Because the lruList is implemented as a Map (or PseudoMap, which
behaves the same on all supported platforms), we can very easily get a
list of all of the keys in the lruList, and then iterate over those,
bypassing any gaps in lu indexes.

This means that the specific fact that mru index is greater than lru
index is no longer important.  Since the numbers are now just more or
less opaque tokens, this replaces the `this._mru++` approach with an
incrementer that will increase the value by one, unless it's
Number.MAX_SAFE_INTEGER, in which case it'll return to 0, and keep doing
this as long as it collides with any lru value in use.

As a side bonus, while doing this, I think I tracked down a CPU-spike
"bug" that had perplexed me a few years back!  (In "scare quotes"
because it was behaving properly, just in a way that was unpleasantly
costly.)

If you have a cache where some items are used much more frequently than
others (pretty much the ideal use-case for an LRU like this one), and
the size of the cache is large enough that some relatively rarely-used
items stick around at the bottom of the barrel, then the distance
between the index of the item you just used, and the oldest item in the
cache, could get very large.  As the top items get used again and again,
the mru counter keeps going up, but the older items never get bumped up,
and the lru value stays low.

This presents a problem when a new item is added, and now we have to
trim the values to be below the max size.  Previously, having no simple
way to iterate by the numeric value of the keys, the only thing to do
was to start at the lru value, and increment up to either the mru value
or the point where the total cache size fell below the maximum.

However, if you have a large hole in that index, it can take a
considerable amount of time to crawl up looking for values, doing
bazillions of object lookups and spinning on CPU forever not responding
to requests.

Fix #50, and hopefully improve performance in at least a few cases that
were otherwise very difficult to debug.