[BUGFIX] offset problems for dragging and scaling with a lot of columns
When using a a lot of columns, e.g. more than 40
then the dashed box showing the next placement would
behave strangely compared to the box I am currently
dragging. This behavior increases while dragging
the box to the right side.

When scaling a box on the right side, the box would
jump one column to the left, which would then move around
all the boxes that I have already placed.

This behavior seems to be related to a rounding problem.