iterating on this further, now doing features without perf impact
Experimented with a few ways to only take the perf hit when
dispose/ttl/size are actually used, and still go fast when they're not.
This seems to work well.

Instead of separate classes for stack and pointer list, it's all just in
the one main LRUCache class.  Then no-op functions are swapped out for
their slower counterparts if features are opted into.