even more iteration protection
There's really no way around checking if the iteration index is valid.
This makes me a little bit sad, because testing for index validity is
slow af. But I suppose cache.entries() and friends are not really likely
to be used in performance-critical paths as much as set() and get().

Re: https://github.com/isaacs/node-lru-cache/issues/212