fix: do not dump() background fetches, track start
If saving a `dump()` to a file or database, and then expecting to load
it again at a later time in another cache object with `cache.load()`
then two awful things could happen.

1. If a background fetch was in progress, it would dump the `Promise`
object in the array.  This is _never_ what you want, and leads to errors
being thrown when attempting to serialize to JSON.

2. The TTL was being dumped, but the _start_ time was not, meaning that
you could dump() an entry with only 1ms left on its 1-hour TTL, but when
reloading it into a new cache, it would be assumed valid for a full
hour.

3. Dumps did not include stale entries at all, so there was no way for a
cache to decide whether to allowStale or not later on.  Sometimes, this
is preferrable.  Now, dump() will always output stale entries (since
they can be resolved accurately for staleness later anyway).