Fail fetch() promises when they are aborted
Fix: https://github.com/isaacs/node-lru-cache/issues/274

Note that this means that some `cache.fetch()` calls that would
previously have succeeded (albeit by returning some invalid value) will
now reject their promises instead.

This is the correct behavior, and so I'm considering this a bugfix/patch
level change, but it does potentially impose some hazards.

The other potentially more subtle difference here is that when a fetch()
is aborted, the promise returned by fetch() will reject *right away*
whereas previously it would still wait for the underlying
fetchMethod() promise to resolve or reject.  This means less time spent
waiting for fetches that aren't ever going to be cached, but it also
means that a fetchMethod that ignores the abort signal could have
previously still returned a value even if it was evicted in the
meantime.

Open to adding an `ignoreFetchAbort` option or something, so that the
previous behavior (resolve to the fetchMethod's returned value, even if
it is no longer covered by the cache) could be maintained, but it's kind
of weird.  If it's aborted, you should abort.  In that case, we'd
probably just not pass a signal to the fetchMethod at all, since
anything that has been touched by an aborted AbortSignal is suspect
data.