Guard against expiration list going missing
This is very mysterious and I have no idea why this is happening, it
seems like it should be impossible, which since it's happening, means
there's a problem in the "seeming".  The expiration list is always
either set to an array, or deleted from the object, which should mean
that any key that exists on the expirations object at any given time is
an array, and there's no re-entry that I can see where it could change
between the start of the loop (where the key has to be in the collection
of keys) and where it's expanded into an array of keys. So, clearly,
either I'm missing something, or the reported issue involves some other
mutation from outside this library.

Probably would be a good idea to port this library to TypeScript and see
where the compiler complains.

Fix: https://github.com/isaacs/ttlcache/issues/26